Added functional save button to save boards.

+++ stky/app/api/boards/[id]/route.js
@@ -6,6 +6,10 @@ export async function GET(request, { params }) {
     try {
         const id = (await params).id;
 
+        if (isNaN(id)) {
+            return NextResponse.json({}, { status: 404 });
+        }
+
         const db = await createConnection();
         const boardsql = `select * from boards
             where id = ${id}`;
@@ -46,6 +50,10 @@ export async function PATCH(request, { params }) {
 
         const id = (await params).id;
 
+        if (isNaN(id)) {
+            return NextResponse.json({}, { status: 404 });
+        }
+
         const checkBoard = `select * from boards
             where id = ${id}`;
 
@@ -65,7 +73,10 @@ export async function PATCH(request, { params }) {
 
         for (let n in body.notes) {
             if (!body.notes[n].id) {
-                return NextResponse.json({error:`Note '${n}' is missing an id.`},{status:422})
+                return NextResponse.json(
+                    { error: `Note '${n}' is missing an id.` },
+                    { status: 422 }
+                );
             }
         }
 
@@ -80,9 +91,15 @@ export async function PATCH(request, { params }) {
             let note = body.notes[n];
 
             if (body.notesOrder.includes(n)) {
-                notessql = `UPDATE notes SET title = '${note.title}', text = '${note.text}', color = ${note.color}, localId = '${n}' 
+                if (note.id == "new") {
+                    notessql = `insert into notes (title,text,boardId,color,localId)
+                values ("${note.title}","${note.text}",${id},${note.color},"${n}");`;
+                } else {
+                    notessql = `UPDATE notes SET title = '${note.title}', text = '${note.text}', color = ${note.color}, localId = '${n}' 
                     WHERE (id = '${note.id}' AND boardId = ${id});`;
+                }
             } else {
+                if(note.id == "new"){continue}
                 notessql = `DELETE FROM notes WHERE (id = ${note.id} AND boardId = ${id});`;
             }
 
@@ -96,8 +113,8 @@ export async function PATCH(request, { params }) {
     }
 }
 
-export async function DELETE(request, {params}){
-    try{
+export async function DELETE(request, { params }) {
+    try {
         const db = await createConnection();
 
         const id = (await params).id;
@@ -111,14 +128,14 @@ export async function DELETE(request, {params}){
             return NextResponse.json({}, { status: 404 });
         }
 
-        const boardsql = `DELETE FROM boards WHERE (id = ${id})`
-        const notessql = `DELETE FROM notes WHERE (boardId = ${id})`
+        const boardsql = `DELETE FROM boards WHERE (id = ${id})`;
+        const notessql = `DELETE FROM notes WHERE (boardId = ${id})`;
 
-        const sendBoard = await db.query(boardsql)
-        const sendNotes = await db.query(notessql)
+        const sendBoard = await db.query(boardsql);
+        const sendNotes = await db.query(notessql);
 
-        return NextResponse.json({})
-    } catch{
+        return NextResponse.json({});
+    } catch {
         console.log(err);
         return NextResponse.json({ error: err.message }, { status: 500 });
     }
